Move issue-detail-drawer out of affected-applications
