+++ client/src/app/pages/issues/affected-applications/affected-applications.tsx
@@ -33,17 +33,13 @@ import {
 } from "@app/shared/components/table-controls";
 import { useFetchAppReports } from "@app/queries/issues";
 import { Link, useLocation, useParams } from "react-router-dom";
-import { IssueFilterGroups } from "../issues";
-import {
-  FilterToolbar,
-  FilterType,
-} from "@app/shared/components/FilterToolbar";
+import { FilterToolbar } from "@app/shared/components/FilterToolbar";
 import { useSelectionState } from "@migtools/lib-ui";
 import {
   getBackToAllIssuesUrl,
   useSharedFilterCategoriesForIssuesAndAffectedApps,
 } from "../helpers";
-import { IssueDetailDrawer } from "./issue-detail-drawer";
+import { IssueDetailDrawer } from "../issue-detail-drawer";
 import { TableURLParamKeyPrefix } from "@app/Constants";
 
 interface IAffectedApplicationsRouteParams {
