+++ client/src/app/pages/issues/issue-detail-drawer/file-incidents-detail-modal/index.ts
