+++ client/src/app/pages/issues/issue-detail-drawer/index.ts
