+++ client/src/app/pages/issues/issue-detail-drawer/issue-affected-files-table.tsx
