+++ client/src/app/pages/issues/issue-detail-drawer/issue-detail-drawer.tsx
